Allow GIMP to upgrade itself
In some rare instances, a file will be replaced with an identically named folder.
The added patch allows eopkg to semi-gracefully handle this by catching the IOError
exception that is raised by the 'tarfile' library and carrying on as if it were a
normal OSError for an already existing file.